Create getIsoluminanceValues.sce
This scenario has been added to the CTraining.exp experiment file, for smoother handling. It precedes CTraining.
It uses the minimal flicker paradigm to match the luminance of green and red stimuli to a light grey and a dark grey background. The light grey background is 10% brighter than the one used throughout the experiment (which is at rgb [82,82,82]), and the dark grey background is 10% darker.
Both the red and green stimuli are presented once at a start value that is much too low, and once too high, for a total of 2*2*2 (colours * starting brightness * background brightness) rounds. It averages the values for the bright and dark starting brightnesses, and writes them to a log file which is later read by the CTraining script.